+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 - A lot of code style warnings (e.g. legacy java 6 code updated, etc)
 ### Added
 - TestItemTree.TestItemLeaf now supports attribute storing
+### Changed
+- Properties loader now ignores property case
 
 ## [5.0.12]
 ### Fixed